Fixed Aug Battery display

+++ DoomRPG/scripts/Utils.c
@@ -1728,7 +1728,7 @@ void DrawBattery()
     SetHudSize(Width, Height, false);
     PrintSpriteFade("AUGBATT", BATTERY_ID, X + 0.4, Y + 0.4, HoldTime, FadeTime);
     SetFont("BIGFONT");
-    HudMessage("%d%%", Player.Augs.Battery);
+    HudMessage("%d%%", (int)Player.Augs.Battery);
     EndHudMessage(HUDMSG_FADEOUT, BATTERY_ID + 1, "Yellow", X + 24.0, Y - 10.0, HoldTime, FadeTime);
 }
